# Conflicts:
#	PlexHelpers.MovieDeDuper/Program.cs
#	PlexSubFix.sln

+++ PlexHelpers.Common/Models/PlexMovie.cs
@@ -9,7 +9,7 @@ namespace PlexHelpers.Common.Models
         public string Guid { get; set; }
         public string CompareTitle { get; set; }
         public int? Year { get; set; }
-        public int Size { get; set; }
+        public long Size { get; set; }
         public int Width { get; set; }
         public int Height { get; set; }
         public decimal DisplayAspectRatio { get; set; }
@@ -24,7 +24,7 @@ namespace PlexHelpers.Common.Models
         public string Hash { get; set; }
         public string MovieFolderName { get; set; }
         public int MetadataId { get; set; }
-        public int Duration { get; set; }
+        public long Duration { get; set; }
         public string IMDB { get; set; }
         public string TMDB { get; set; }
         public string TVDB { get; set; }
